feat: unifie le hook de gestion d'entité
