+++ src/entities/core/hooks/index.ts
@@ -1,10 +1,9 @@
 "use client";
 export { default as useEntityManager } from "./useEntityManager";
-export { default as useEntityManagerGeneral } from "./useEntityManagerGeneral";
 export type {
     FieldKey,
     SingleFieldConfig,
     FieldConfig,
     UseEntityManagerOptions,
     EntityManagerResult,
-} from "./useEntityManagerGeneral";
+} from "./useEntityManager";
