+++ src/entities/core/hooks/useEntityManager.ts
@@ -4,17 +4,27 @@ import { useEffect, useState, useCallback } from "react";
 
 export type FieldKey<T> = keyof T & string;
 
-export interface UseEntityManagerOptions<T extends Record<string, string>> {
+export interface SingleFieldConfig<V> {
+    parse: (input: string) => V;
+    serialize: (value: V) => V;
+    validate?: (value: V) => boolean;
+    emptyValue: V;
+}
+
+export type FieldConfig<T> = { [K in FieldKey<T>]: SingleFieldConfig<T[K]> };
+
+export interface UseEntityManagerOptions<T extends Record<string, unknown>> {
     fetch: () => Promise<(T & { id?: string }) | null>;
     create: (data: T) => Promise<void>;
-    update: (entity: (T & { id?: string }) | null, data: Record<string, string>) => Promise<void>;
+    update: (entity: (T & { id?: string }) | null, data: Partial<T>) => Promise<void>;
     remove: (entity: (T & { id?: string }) | null) => Promise<void>;
     labels: (field: FieldKey<T>) => string;
     fields: FieldKey<T>[];
     initialData: T;
+    config: FieldConfig<T>;
 }
 
-export interface EntityManagerResult<T extends Record<string, string>> {
+export interface EntityManagerResult<T extends Record<string, unknown>> {
     entity: (T & { id?: string }) | null;
     formData: T;
     setFormData: React.Dispatch<React.SetStateAction<T>>;
@@ -35,7 +45,7 @@ export interface EntityManagerResult<T extends Record<string, string>> {
     fetchData: () => Promise<void>;
 }
 
-export default function useEntityManager<T extends Record<string, string>>({
+export default function useEntityManager<T extends Record<string, unknown>>({
     fetch,
     create,
     update,
@@ -43,6 +53,7 @@ export default function useEntityManager<T extends Record<string, string>>({
     labels,
     fields,
     initialData,
+    config,
 }: UseEntityManagerOptions<T>): EntityManagerResult<T> {
     const [entity, setEntity] = useState<(T & { id?: string }) | null>(null);
     const [formData, setFormData] = useState<T>(initialData);
@@ -60,9 +71,9 @@ export default function useEntityManager<T extends Record<string, string>>({
             const data = await fetch();
             setEntity(data);
             if (data) {
-                const next = { ...initialData };
+                const next = { ...initialData } as T;
                 fields.forEach((f) => {
-                    next[f] = data[f] ?? "";
+                    next[f] = data[f];
                 });
                 setFormData(next);
             }
@@ -82,22 +93,26 @@ export default function useEntityManager<T extends Record<string, string>>({
     // ...le reste inchangé
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         const { name, value } = e.target;
-        setFormData((f) => ({ ...f, [name]: value }));
+        const field = name as FieldKey<T>;
+        const parsed = config[field].parse(value);
+        if (config[field].validate && !config[field].validate(parsed)) return;
+        setFormData((f) => ({ ...f, [field]: parsed }));
     };
 
     const save = async () => {
         setLoading(true);
         try {
+            const serialized = { ...formData } as T;
+            fields.forEach((f) => {
+                serialized[f] = config[f].serialize(formData[f]);
+            });
             if (entity) {
-                await update(entity, formData);
+                await update(entity, serialized);
                 await fetchData();
             } else {
-                await create(formData);
+                await create(serialized);
                 await fetchData();
             }
-
-            // Optionnel : refresh après save si besoin
-            // await fetchData();
         } finally {
             setLoading(false);
             setEditMode(false);
@@ -109,11 +124,12 @@ export default function useEntityManager<T extends Record<string, string>>({
         setLoading(true);
         try {
             const { field, value } = editModeField;
-            await update(entity, { [field]: value });
-            setFormData((f) => ({ ...f, [field]: value }));
+            const parsed = config[field].parse(value);
+            if (config[field].validate && !config[field].validate(parsed)) return;
+            const serialized = config[field].serialize(parsed);
+            await update(entity, { [field]: serialized } as Partial<T>);
+            setFormData((f) => ({ ...f, [field]: parsed }));
             setEditModeField(null);
-            // Optionnel : refresh après saveField si besoin
-            // await fetchData();
         } finally {
             setLoading(false);
         }
@@ -122,10 +138,10 @@ export default function useEntityManager<T extends Record<string, string>>({
     const clearField = async (field: FieldKey<T>) => {
         setLoading(true);
         try {
-            await update(entity, { [field]: "" });
-            setFormData((f) => ({ ...f, [field]: "" }));
-            // Optionnel : refresh après clearField si besoin
-            // await fetchData();
+            const empty = config[field].emptyValue;
+            const serialized = config[field].serialize(empty);
+            await update(entity, { [field]: serialized } as Partial<T>);
+            setFormData((f) => ({ ...f, [field]: empty }));
         } finally {
             setLoading(false);
         }
