+++ src/entities/core/index.ts
@@ -1,3 +1,4 @@
 export * from "./services";
 export * from "./utils";
 export * from "./types";
+export * from "./hooks";
